V3_ADD-EMAIL

+++ src/main/kotlin/com/proyectoxavier/factura/model/Detail.kt
@@ -10,8 +10,8 @@ class Detail {
     @Id
     @Column(updatable = false)
     var id: Long? = null
-    var quantity: Int? = null
-    var price: Int? = null
+    var quantity: Long? = null
+    var price: Long? = null
     @Column(name="invoice_id")
     var invoiceId: Long? = null
     @Column(name="product_id")
